packages/gtmcore/requirements.txt
pyyaml==3.12
GitPython==3.0.8
gitdb-speedups==0.1.0
docker==3.4.1
redis==3.3.0
hiredis==1.0.0
python-redis-lock==3.2.0
rq==0.13
rq-scheduler==0.9
jupyter_client==5.2.3
ipykernel==4.8.2
requests==2.20.0
cryptography==2.3.1
python-jose[cryptography]==3.0.0
schema==0.6.6
blosc==1.5.1
pexpect==4.3.1
natsort==5.1.1
pillow==6.2.1
# click conflicts  7.0 installed by flask-api, 6.7 OK for both
click==6.7
mitmproxy==4.0.4
pandas==0.23.4
aiohttp==3.5.4
aiofiles==0.4.0

# Dataset backend specific
python-snappy==0.5.3
boto3==1.9.103
packaging==19.0
humanfriendly==4.18
